practice on thread Example
