+++ SpringBootCodexBoxJava2022Batch1/src/main/java/com/codexbox/springboot/app/vidyadarna/threadconcepts/ThreadExample.java
@@ -1,0 +1,5 @@
+package com.codexbox.springboot.app.vidyadarna.threadconcepts;
+
+public class ThreadExample extends Thread{
+
+}
